rename test files to remove v2
