+++ packages/language-support/src/tests/formatting/comments.test.ts
@@ -4,7 +4,7 @@
  * since it would be difficult to consolidate the new and the old version
  */
 
-import { verifyFormatting } from './testutilv2';
+import { verifyFormatting } from './testutil';
 
 describe('should not forget to include all comments', () => {
   test('property comments', () => {
